requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.2.0
django-import-export==2.1.0
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.9
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.4
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
